backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pyjwt
httpx
python-multipart
python-dotenv
passlib[bcrypt]
pydantic-settings
bcrypt
supabase
prometheus-fastapi-instrumentator
sentry-sdk[fastapi]
requests
torch
torchvision
torchaudio
transformers
trl
accelerate
datasets
peft
gymnasium
stable-baselines3
psycopg2-binary